requirements.txt
# Telegram Bot
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0

# Mini App Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# Charts and Data
matplotlib==3.8.2
seaborn==0.13.1

# Database
# SQLite is built-in with Python

# Utilities
requests==2.31.0
